Reference archival URL
The PSF Trademarks Working Group was notified that the pycon.dk page is being cybersquatted by a commercial sales site with no relation to the Python programming language.  We suggest that referencing the last "authentic" version of the page is a good approach.  The underlying organizers may still be active, but might simply have let domain lapse.