Update online 2020-21 and announce future plans
Not related to Trademark Committee per se, but since I took a closer look, it seems like we can update the last few years, and next few, in light of the pandemic-driven changes.

An ampersand and the word 'and' were intermixed in the prior paragraph.  I have made it consistently use the word instead.