Update URLs to use https instead of https
This is a bulk update to resolve almost all of the htmlproofer CI errors
(and generally to link to secure versions of websites). The remaining 7
errors reported by htmlproofer are conference-specific and relate to the
inability to validate links via libcurl, which may be due to servers not
supporting https or domains no longer resolving.